requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
redis>=5.0
gunicorn>=21.0
django-environ>=0.11
whitenoise>=6.5
weasyprint>=61.0
django-weasyprint>=2.2
djangorestframework>=3.14
django-cors-headers>=4.3
